users_service/requirements.txt
Django==4.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
psycopg2-binary==2.9.6
python-decouple==3.8
django-cors-headers==4.0.0
requests==2.31.0
PyJWT==2.6.0 